Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/script-tag-post-control-char-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?test=/security/xssAuditor/script-tag-post-control-char.html' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
